+++ sdk/spring/spring-cloud-azure-appconfiguration-config/src/main/java/com/azure/spring/cloud/appconfiguration/config/implementation/AppConfigurationPullRefresh.java
@@ -39,6 +39,8 @@ public class AppConfigurationPullRefresh implements AppConfigurationRefresh {
     private final Duration refreshInterval;
     
     private final ReplicaLookUp replicaLookUp;
+    
+    private final AppConfigurationRefreshUtil refreshUtils;
 
     /**
      * Component used for checking for and triggering configuration refreshes.
@@ -46,13 +48,16 @@ public class AppConfigurationPullRefresh implements AppConfigurationRefresh {
      * @param clientFactory Clients stores used to connect to App Configuration. * @param defaultMinBackoff default
      * @param refreshInterval time between refresh intervals
      * @param defaultMinBackoff minimum time between backoff retries minimum backoff time
+     * @param replicaLookUp looks up new replicas that can be used
+     * @param refreshUtils utility for dealing with refresh
      */
     public AppConfigurationPullRefresh(AppConfigurationReplicaClientFactory clientFactory, Duration refreshInterval,
-        Long defaultMinBackoff, ReplicaLookUp replicaLookUp) {
+        Long defaultMinBackoff, ReplicaLookUp replicaLookUp, AppConfigurationRefreshUtil refreshUtils) {
         this.defaultMinBackoff = defaultMinBackoff;
         this.refreshInterval = refreshInterval;
         this.clientFactory = clientFactory;
         this.replicaLookUp = replicaLookUp;
+        this.refreshUtils = refreshUtils;
 
     }
 
@@ -70,6 +75,17 @@ public class AppConfigurationPullRefresh implements AppConfigurationRefresh {
      */
     public Mono<Boolean> refreshConfigurations() {
         return Mono.just(refreshStores());
+    }
+    
+    /**
+     * Checks configurations to see if configurations should be reloaded. If the refresh interval has passed and a
+     * trigger has been updated configuration are reloaded.
+     *
+     * @return Mono with a boolean of if a RefreshEvent was published. If refreshConfigurations is currently being run
+     * elsewhere this method will return right away as <b>false</b>.
+     */
+    public void refreshAsync() {
+        new Thread(new RunnableRefresh()).start();
     }
 
     /**
@@ -98,7 +114,7 @@ public class AppConfigurationPullRefresh implements AppConfigurationRefresh {
         if (running.compareAndSet(false, true)) {
             BaseAppConfigurationPolicy.setWatchRequests(true);
             try {
-                RefreshEventData eventData = AppConfigurationRefreshUtil.refreshStoresCheck(clientFactory,
+                RefreshEventData eventData = refreshUtils.refreshStoresCheck(clientFactory,
                     refreshInterval, defaultMinBackoff, replicaLookUp);
                 if (eventData.getDoRefresh()) {
                     publisher.publishEvent(new RefreshEvent(this, eventData, eventData.getMessage()));
@@ -118,6 +134,18 @@ public class AppConfigurationPullRefresh implements AppConfigurationRefresh {
     @Override
     public Map<String, AppConfigurationStoreHealth> getAppConfigurationStoresHealth() {
         return clientFactory.getHealth();
+    }
+    
+    /**
+     * Runnable refresh. Created so Async only application can use refresh, such as WebFlux.
+     */
+    private class RunnableRefresh implements Runnable {
+
+        @Override
+        public void run() {
+            refreshStores();
+        }
+        
     }
 
 }